// src/server/http/inovopay/inovopay.schema.ts
export namespace InovoPay {
// ...
  export const basicAuthSchema = z.object({
    REQUEST_ACTION: z.string(),
    REQ_USERNAME: z.string(),
    REQ_PASSWORD: z.string(),
    REQUEST_RESPONSE_FORMAT: z.enum(['XML', 'PIPES', 'JSON']).default('JSON'),
    REQUEST_API_VERSION: z.string(),
    SITE_ID: z.number(),
  });
  export type BasicAuth = z.infer<typeof basicAuthSchema>;

  export const creditCardTransactionInputSchema = z.object({
    CUST_FNAME: z.string().nullish(),
    CUST_LNAME: z.string().nullish(),
    CUST_EMAIL: z.string().nullish(),
    LI_COUNT_1: z.number(),
    LI_PROD_ID_1: z.number(),
    LI_VALUE_1: z.number(),
    XTL_ORDER_ID: z.string().nullish(),
    BILL_ADDR: z.string().nullish(),
    BILL_ADDR_CITY: z.string().nullish(),
    BILL_ADDR_STATE: z.string().nullish(),
    BILL_ADDR_ZIP: z.string().nullish(),
    BILL_ADDR_COUNTRY: z.string().nullish(),
    PMT_NUMB: z.string(),
    TOKEN_GUID: z.string().nullish(),
    PMT_KEY: z.string(),
    PMT_EXPIRY: z.string(),
    CUST_LOGIN: z.string().nullish(),
    CUST_PASSWORD: z.string().nullish(),
    MERCH_ACCT_ID: z.string().nullish(),
    REQUEST_CURRENCY: z.string().default('USD'),
    PMT_DESCRIPTOR: z.string().nullish(),
    PMT_DESCRIPTOR_PHONE: z.string().nullish(),
    CUST_PHONE: z.string().nullish(),
    REQUEST_AFF_ID: z.string().nullish(),
    REQUEST_AFF_ID_SUB: z.string().nullish(),
    UNIQUE_XTL_ORDER_ID: z.enum(['0', '1', '2']).nullish(),
    PMT_ID_XTL: z.string().nullish(),
    MBSHIP_ID_XTL: z.string().nullish(),
    TRANS_REBILL_TYPE: z.enum(['NONE', 'TRIAL', 'INITIAL', 'REBILL']).default('NONE'),
    TRANS_TRIAL_REBILL_CUSTOMER_CONSENT: z.string().nullish(),
    TRANS_CUSTOMER_RECEIPT: z.string().nullish(),
    CARD_ON_FILE_FLAG: z.enum(['0', '1']),
  });

  export type CreditCardTransactionInput = z.infer<typeof creditCardTransactionInputSchema>;

  export const creditCardTransactionResponseSchema = z.object({
    REQUEST_ACTION: z.string().nullish(),
    TRANS_STATUS_NAME: z.string().nullish(),
    TRANS_VALUE: z.string().nullish(),
    TRANS_VALUE_SETTLED: z.string().nullish(),
    CURR_CODE_ALPHA_SETTLED: z.string().nullish(),
    TRANS_EXCH_RATE: z.string().nullish(),
    TRANS_ID: z.string().nullish(),
    CUST_ID: z.string().nullish(),
    XTL_CUST_ID: z.string().nullish(),
    REQ_ID: z.string().nullish(),
    PO_ID: z.string().nullish(),
    XTL_ORDER_ID: z.string().nullish(),
    BATCH_ID: z.string().nullish(),
    PROC_NAME: z.string().nullish(),
    MERCH_ACCT_ID: z.string().nullish(),
    CARD_BRAND_NAME: z.string().nullish(),
    CARD_DETAIL: z.string().nullish(),
    CARD_TYPE: z.string().nullish(),
    PMT_AAU_UPDATE_DT: z.string().nullish(),
    PMT_AAU_UPDATE_DESC: z.string().nullish(),
    CARD_PREPAID: z.string().nullish(),
    CARD_BANK: z.string().nullish(),
    CARD_BALANCE: z.string().nullish(),
    PMT_L4: z.string().nullish(),
    PMT_ID: z.string().nullish(),
    PMT_ID_XTL: z.string().nullish(),
    PROC_UDF01: z.string().nullish(),
    PROC_UDF02: z.string().nullish(),
    PROC_AUTH_RESPONSE: z.string().nullish(),
    PROC_RETRIEVAL_NUM: z.string().nullish(),
    PROC_REFERENCE_NUM: z.string().nullish(),
    PROC_REDIRECT_URL: z.string().nullish(),
    AVS_RESPONSE: z.string().nullish(),
    CVV_RESPONSE: z.string().nullish(),
    REQUEST_API_VERSION: z.string().nullish(),
    PO_LI_ID_1: z.number().nullish(),
    PO_LI_COUNT_1: z.number().nullish(),
    PO_LI_AMOUNT_1: z.number().nullish(),
    PO_LI_PROD_ID_1: z.number().nullish(),
    MBSHIP_ID_1: z.string().nullish(),
    // Error/decline/partial fields
    API_RESPONSE: z.string().nullish(),
    API_ADVICE: z.string().nullish(),
    SERVICE_RESPONSE: z.union([z.string(), z.number()]).nullish(),
    SERVICE_ADVICE: z.string().nullish(),
    PROCESSOR_RESPONSE: z.union([z.string(), z.number()]).nullish(),
    PROCESSOR_ADVICE: z.string().nullish(),
    INDUSTRY_RESPONSE: z.union([z.string(), z.number()]).nullish(),
    INDUSTRY_ADVICE: z.string().nullish(),
    REF_FIELD: z.string().nullish(),
  });
  export type CreditCardTransactionResponse = z.infer<typeof creditCardTransactionResponseSchema>;
}